# api/openapi.yaml
openapi: 3.0.0
info:
  title: FUSKAMO Backend API
  version: 1.0.0
  description: >
    Custom backend fronting M-Pesa, AI services, external APIs, notifications,
    and admin operations. See backend/README.md and docs/admin-access.md.
    Two distinct auth schemes are in use — see securitySchemes below; the
    admin ones use real per-admin Supabase Auth sessions (JWT bearer tokens),
    never a shared key.
servers:
  - url: http://localhost:8080/api
paths:
...
  /admin/transactions:
    get:
      summary: Recent M-Pesa transactions
      security: [{ AdminAuth: [] }]
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
      responses:
        '200': { description: List of transactions }
        '403': { description: Admin lacks the payments:view permission }
...
  /admin-accounts:
    get:
      summary: List all administrator accounts
      security: [{ AdminAuth: [] }]
      responses:
        '200': { description: List of admins }
        '403': { description: Admin lacks the admins:view permission }
...
  /search:
    get:
      summary: Public universal search across approved players, verified scouts, and public groups. Used by the web-app's global search box.
      parameters:
        - { name: q, in: query, required: true, schema: { type: string } }
      responses:
        '200': { description: Matching results across players, scouts, and groups }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: >
        Shared secret between the Flutter app and this backend
        (BACKEND_API_KEY). Not user identity — just stops unauthenticated
        internet traffic from hitting M-Pesa/AI endpoints. See
        authentication/apiKeyAuth.js.
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        A real Supabase Auth session belonging to an administrator with a
        row in admin_users. NOT the same as ApiKeyAuth above, and NOT a
        shared secret of any kind — every admin has their own account, own
        role, and own granular permissions (see docs/admin-access.md and
        authentication/adminAuth.js). Requests also need the specific
        permission each route declares (authorization/permissions.js);
        having a valid admin session alone is not sufficient for every
        admin route.
    ScoutAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        A real Supabase Auth session belonging to any signed-in user — NOT
        an admin session (see AdminAuth) and NOT the shared ApiKeyAuth
        secret. Used only by /discovery/*, which needs to know which scout
        is asking so it can look up their preferences and attribute
        engagement events (authentication/jwtAuth.js). The endpoint itself
        further requires a verified row in scouts linked via user_id —
        a valid session with no verified scout profile gets a 404, not a
        403.